New version with more controlle

+++ crm-api/app/Http/Controllers/InvitationController.php
@@ -5,41 +5,33 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Cache;
+
+use App\Http\Requests\InvitationRequest;
 
 use App\Models\Invitation;
 use App\Models\Log;
 
 use App\Mail\InvitationMail;
 
+use App\Events\LogGenerate;
+
 class InvitationController extends Controller
 {
     /**
      * Valid and create Invitation
      */
-    public function create(Request $request){ 
-        $validate = Validator::make($request->all(),[
-            'name' => 'string|required',
-            'email' => 'email|required',
-            'companie' => 'integer|required', 
-        ]);
-        if($validate->fails())
-            return response()->render('error', $validate->messages()->all());
+    public function create(InvitationRequest $request){  
         $send = $request->get('status');
         
-        $invitation = $request->all();
+        $invitation = $request->validated();
 
         $invitation = Invitation::create($invitation);
         if($invitation){
             //Check if this invitation is sended with send email request 
             if($send == 'sent'){
-                Mail::to($invitation->email)->send(new InvitationMail($invitation));
-                Log::create([
-                    'action' => 'create',
-                    'model' => 'invitation',
-                    'message' => 'Admin '.Auth()->user()->name.' create invitation to email '.$invitation->email.' with '.$send.' status',
-                    'target' => $invitation->id,
-                    'trigger' => Auth()->user()->id,
-                ]);
+                Mail::to($invitation->email)->send(new InvitationMail($invitation)); 
+                LogGenerate::dispatch($invitation, 'create', 'Admin '.Auth()->user()->name.' create invitation to email '.$invitation->email.' with '.$send.' status');
             }
             return $this->get('Invitation successfully created you can create more');
         }
@@ -48,11 +40,15 @@ class InvitationController extends Controller
     }
 
     /**
-     * List all companies
+     * List all invitations
      */
-    public function get($message = 'Companies successfully listed')
+    public function get($message = 'Invitations successfully listed')
     {
-        return response()->render('success', $message, Invitation::latest()->get());
+        $invitations = Cache::remember('invitations', 120, function () {
+            return collect(Invitation::cursor());
+        });
+
+        return response()->render('success', $message, $invitations);
     }
 
     /**
@@ -60,17 +56,13 @@ class InvitationController extends Controller
      */   
 
     public function delete(Invitation $invitation){
+        if($invitation->status == 'validated')
+            return response()->render('error', 'Invitation already validated');
 
         $success = $invitation->delete();
 
-        if($success)
-            Log::create([
-                'action' => 'delete',
-                'model' => 'invitation',
-                'message' => 'Admin '.Auth()->user()->name.' delete invitation '.$invitation->email,
-                'target' => $invitation->id,
-                'trigger' => Auth()->user()->id,
-            ]);
+        if($success) 
+            LogGenerate::dispatch($invitation, 'delete', 'Admin '.Auth()->user()->name.' delete invitation '.$invitation->email);
         return $this->get('Invitation successfully deleted');
     }
 
@@ -80,5 +72,14 @@ class InvitationController extends Controller
 
     public function getById(Invitation $invitation){
         return response()->render('success','Invitation successfully returned', $invitation);
+    }
+
+    public function sendEmail(Invitation $invitation){
+        if($invitation->status == 'validated')
+            return response()->render('error', 'Invitation already sent');
+
+        Mail::to($invitation->email)->send(new InvitationMail($invitation));
+
+        return $this->get('Invitation successfully sent');
     }
 }
